don't resize pane with delta
say the pane with 100 cols wide, we got 3 mouse events, one move the
split to 99, the next one move it to 98, then to 90

we would first call resize_split_by with -1, -1, then with -8. the first
resize is completed just as we call resize_split_by(-8), this triggers a
resync, and updates the split position to 99. because of that
resize_split_by(-8) will try to resize it to 91, instead of the desire
90.

in practice, there are a lot more mouse events and the timing is more
unpredictable, the end result can be very chaotic.

partially fix wez#5142